+++ confluence/legacy-mobile-macros/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/legacy-mobile-macros
 
+## 2.1.19
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 2.1.18
 
 ### Patch Changes