+++ drag-and-drop/autoscroll/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/drag-and-drop-autoscroll
 
+## 0.1.9
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 0.1.8
 
 ### Patch Changes