+++ drag-and-drop/hitbox/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/drag-and-drop-hitbox
 
+## 0.2.7
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 0.2.6
 
 ### Patch Changes