+++ editor/editor-common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/editor-common
 
+## 70.1.2
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 70.1.1
 
 ### Patch Changes